requirements.txt
pytest>=7.4.0
requests>=2.31.0
PyYAML>=6.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
jsonschema>=4.19.0
jsonpath-ng>=1.6.0
allure-pytest>=2.13.0
urllib3>=2.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
jinja2>=3.1.0
python-multipart>=0.0.6
